requirements.txt
discord
jishaku
discord-pretty-help
qrcode
pillow
pyshorteners
bhimupi
binlist